# ⚡️ Speed up function `get_1d_sincos_pos_embed_from_grid_torch` by 16%