### 📄 16% (0.16x) speedup for ***`get_1d_sincos_pos_embed_from_grid_torch` in `torchgeo/models/panopticon.py`***

⏱️ Runtime :   **`4.15 milliseconds`**  **→** **`3.59 milliseconds`** (best of `163` runs)
### 📝 Explanation and details

Here’s an optimized version of your code. See the comments below for speed-up details.

- **Avoid redundant reshaping**: `pos` is by contract already 1D (enforced by `.reshape(-1)`), but it's harmless and fast.
- **Move all possible operations to a batched form and replace einsum with torch.outer**, which is much faster for this simple outer product.
- **Fuse operations to reduce temporaries**.

Optimized Code.


**Summary of changes**.
- Used `torch.outer` instead of `einsum` (10-30% speedup for this shape).
- Use tuple form in `torch.cat` for minor performance gain.
- Clarified variable names and simplified arithmetic.
- Retained all output shapes and function signature as required.

This is about as fast as you'll get using native PyTorch and the basic formulation!


✅ **Correctness verification report:**

| Test                        | Status            |
| --------------------------- | ----------------- |
| ⚙️ Existing Unit Tests | ✅ **26 Passed** |
| 🌀 Generated Regression Tests | ✅ **35 Passed** |
| ⏪ Replay Tests | 🔘 **None Found** |
| 🔎 Concolic Coverage Tests | 🔘 **None Found** |
|📊 Tests Coverage       | 100.0% |
<details>
<summary>⚙️ Existing Unit Tests and Runtime</summary>

| Test File::Test Function                                                                 | Original ⏱️   | Optimized ⏱️   | Speedup   |
|:-----------------------------------------------------------------------------------------|:--------------|:---------------|:----------|
| `models/test_panopticon.py::TestGet1DSincosPosEmbedFromGridTorch.test_invalid_embed_dim` | 1.24μs        | 1.11μs         | ✅11.9%   |
| `models/test_panopticon.py::TestGet1DSincosPosEmbedFromGridTorch.test_output_shape`      | 106μs         | 91.1μs         | ✅17.0%   |

</details>

<details>
<summary>🌀 Generated Regression Tests and Runtime</summary>

```python
import pytest  # used for our unit tests
import torch
from torch import Tensor
from torchgeo.models.panopticon import get_1d_sincos_pos_embed_from_grid_torch

# unit tests

# ------------------------
# Basic Test Cases
# ------------------------

def test_basic_single_position():
    # Test with embed_dim=4 and a single position
    embed_dim = 4
    pos = torch.tensor([0.], dtype=torch.float32)
    codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos); emb = codeflash_output # 108μs -> 89.4μs (21.8% faster)

def test_basic_multiple_positions():
    # Test with embed_dim=6 and multiple positions
    embed_dim = 6
    pos = torch.tensor([0., 1., 2.], dtype=torch.float32)
    codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos); emb = codeflash_output # 106μs -> 89.9μs (18.5% faster)

def test_basic_dtype_and_device():
    # Test with float64 and on cpu/cuda if available
    embed_dim = 8
    pos = torch.arange(4, dtype=torch.float64)
    codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos.float()); emb = codeflash_output # 92.6μs -> 73.7μs (25.6% faster)
    # If cuda is available, test on cuda
    if torch.cuda.is_available():
        pos_cuda = pos.cuda().float()
        codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos_cuda); emb_cuda = codeflash_output

# ------------------------
# Edge Test Cases
# ------------------------

def test_embed_dim_not_even():
    # Should raise assertion error if embed_dim is not even
    pos = torch.arange(2, dtype=torch.float32)
    with pytest.raises(AssertionError):
        get_1d_sincos_pos_embed_from_grid_torch(3, pos)

def test_empty_positions():
    # Should handle empty pos tensor gracefully
    embed_dim = 6
    pos = torch.tensor([], dtype=torch.float32)
    codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos); emb = codeflash_output # 111μs -> 93.7μs (18.8% faster)

def test_negative_positions():
    # Should handle negative positions correctly (sin/cos are defined for all reals)
    embed_dim = 4
    pos = torch.tensor([-2., -1., 0., 1., 2.], dtype=torch.float32)
    codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos); emb = codeflash_output # 109μs -> 91.0μs (20.5% faster)
    # Check symmetry: sin(-x) == -sin(x), cos(-x) == cos(x)
    emb_neg2 = emb[0]
    emb_pos2 = emb[4]

def test_large_embed_dim():
    # Test with a large embed_dim (close to 1000, but ensure < 100MB)
    embed_dim = 1000
    pos = torch.tensor([0., 1.], dtype=torch.float32)
    codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos); emb = codeflash_output # 113μs -> 94.5μs (20.0% faster)

def test_non_contiguous_positions():
    # Test with a non-contiguous pos tensor (e.g., from slicing)
    embed_dim = 8
    arr = torch.arange(10, dtype=torch.float32)
    pos = arr[::2]  # non-contiguous
    codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos); emb = codeflash_output # 99.7μs -> 82.6μs (20.8% faster)

def test_positions_with_inf_nan():
    # Test with positions containing inf or nan
    embed_dim = 4
    pos = torch.tensor([0., float('inf'), float('nan')], dtype=torch.float32)
    codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos); emb = codeflash_output # 108μs -> 93.1μs (16.0% faster)

def test_positions_with_large_values():
    # Test with very large positions (should not overflow)
    embed_dim = 4
    pos = torch.tensor([1e10, -1e10], dtype=torch.float32)
    codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos); emb = codeflash_output # 107μs -> 89.0μs (21.1% faster)

# ------------------------
# Large Scale Test Cases
# ------------------------

def test_large_number_of_positions():
    # Test with a large number of positions (e.g., 10,000)
    embed_dim = 32
    N = 10000
    pos = torch.arange(N, dtype=torch.float32)
    codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos); emb = codeflash_output # 313μs -> 312μs (0.541% faster)

def test_large_embed_dim_and_positions():
    # Test with both large embed_dim and positions, but keep <100MB
    embed_dim = 512  # 512*1000*4 = 2MB
    N = 1000
    pos = torch.arange(N, dtype=torch.float32)
    codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos); emb = codeflash_output # 355μs -> 336μs (5.61% faster)

def test_gradient_flow():
    # Test that the output is differentiable with respect to pos
    embed_dim = 8
    pos = torch.arange(4, dtype=torch.float32, requires_grad=True)
    codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos); emb = codeflash_output # 120μs -> 101μs (18.1% faster)
    # Take sum and backward
    emb.sum().backward()

def test_output_is_deterministic():
    # Test that repeated calls with same input yield same output
    embed_dim = 16
    pos = torch.linspace(0, 10, 20)
    codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos); emb1 = codeflash_output # 108μs -> 90.4μs (19.9% faster)
    codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos); emb2 = codeflash_output # 40.5μs -> 30.1μs (34.5% faster)
# codeflash_output is used to check that the output of the original code is the same as that of the optimized code.

import pytest  # used for our unit tests
import torch
from torch import Tensor
from torchgeo.models.panopticon import get_1d_sincos_pos_embed_from_grid_torch

# unit tests

# ---------------------- Basic Test Cases ----------------------

def test_basic_single_position():
    # Test with a single position and small embed_dim
    pos = torch.tensor([0.0])
    embed_dim = 4
    codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos); emb = codeflash_output # 107μs -> 89.2μs (20.2% faster)
    # For pos=0, sin(0)=0, cos(0)=1, so first half should be zeros, second half ones
    torch.testing.assert_close(emb[0, :embed_dim//2], torch.zeros(embed_dim//2))
    torch.testing.assert_close(emb[0, embed_dim//2:], torch.ones(embed_dim//2))

def test_basic_multiple_positions():
    # Test with multiple positions and a moderate embed_dim
    pos = torch.tensor([0.0, 1.0, 2.0])
    embed_dim = 6
    codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos); emb = codeflash_output # 109μs -> 92.3μs (19.1% faster)
    # For pos=0, first row should be [0,0,0,1,1,1]
    torch.testing.assert_close(emb[0, :embed_dim//2], torch.zeros(embed_dim//2))
    torch.testing.assert_close(emb[0, embed_dim//2:], torch.ones(embed_dim//2))

def test_basic_non_integer_positions():
    # Test with float positions (should work for float input)
    pos = torch.tensor([0.0, 0.5, 1.5])
    embed_dim = 8
    codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos); emb = codeflash_output # 111μs -> 91.8μs (21.5% faster)

def test_basic_device_consistency():
    # Test that function works on CUDA if available
    if torch.cuda.is_available():
        pos = torch.tensor([0.0, 1.0, 2.0], device='cuda')
        embed_dim = 4
        codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos); emb = codeflash_output

def test_basic_dtype_consistency():
    # Test that function outputs float32 regardless of input dtype
    pos = torch.tensor([0, 1, 2], dtype=torch.int64)
    embed_dim = 4
    codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos); emb = codeflash_output # 117μs -> 98.5μs (19.5% faster)

# ---------------------- Edge Test Cases ----------------------

def test_edge_embed_dim_not_even():
    # Should raise AssertionError if embed_dim is not even
    pos = torch.tensor([0.0, 1.0])
    with pytest.raises(AssertionError):
        get_1d_sincos_pos_embed_from_grid_torch(3, pos)

def test_edge_zero_positions():
    # Test with empty positions tensor
    pos = torch.tensor([])
    embed_dim = 4
    codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos); emb = codeflash_output # 111μs -> 93.3μs (19.7% faster)


def test_edge_negative_positions():
    # Test with negative position values
    pos = torch.tensor([-1.0, -2.0, 0.0, 1.0])
    embed_dim = 4
    codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos); emb = codeflash_output # 116μs -> 98.6μs (17.8% faster)

def test_edge_large_negative_and_positive_positions():
    # Test with very large negative and positive positions
    pos = torch.tensor([-1e6, 0.0, 1e6])
    embed_dim = 8
    codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos); emb = codeflash_output # 109μs -> 93.1μs (17.3% faster)

def test_edge_non_contiguous_input():
    # Test with non-contiguous input tensor
    pos = torch.arange(10)[::2]  # [0,2,4,6,8], non-contiguous
    embed_dim = 4
    codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos); emb = codeflash_output # 105μs -> 87.1μs (21.1% faster)

def test_edge_high_dim_input():
    # Test with input of shape (N,1) instead of (N,)
    pos = torch.tensor([[0.0], [1.0]])
    embed_dim = 4
    codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos); emb = codeflash_output # 105μs -> 90.0μs (17.6% faster)

# ---------------------- Large Scale Test Cases ----------------------

def test_large_many_positions():
    # Test with a large number of positions (but <1000 for memory limit)
    N = 999
    embed_dim = 32
    pos = torch.arange(N, dtype=torch.float32)
    codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos); emb = codeflash_output # 157μs -> 136μs (15.2% faster)

def test_large_high_embed_dim():
    # Test with a large embed_dim (but <1000 for memory limit)
    N = 10
    embed_dim = 998
    pos = torch.arange(N, dtype=torch.float32)
    codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos); emb = codeflash_output # 114μs -> 97.2μs (18.0% faster)

def test_large_positions_and_embed_dim():
    # Test with both large N and large embed_dim, but under 100MB
    N = 100
    embed_dim = 800
    pos = torch.linspace(-1000, 1000, N)
    codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos); emb = codeflash_output # 155μs -> 136μs (14.1% faster)

def test_large_performance():
    # Test that function runs efficiently on large, but manageable input
    N = 500
    embed_dim = 500
    pos = torch.arange(N, dtype=torch.float32)
    codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos); emb = codeflash_output # 230μs -> 216μs (6.29% faster)

# ---------------------- Miscellaneous Robustness ----------------------

def test_mixed_dtype_input():
    # Test with integer positions (should be cast to float internally)
    pos = torch.tensor([1, 2, 3], dtype=torch.int32)
    embed_dim = 4
    codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos); emb = codeflash_output # 113μs -> 96.8μs (17.0% faster)

def test_gradients():
    # Test that the output is differentiable wrt input positions
    pos = torch.tensor([0.5, 1.5, 2.5], requires_grad=True)
    embed_dim = 6
    codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos); emb = codeflash_output # 121μs -> 105μs (15.4% faster)
    loss = emb.sum()
    loss.backward()

def test_output_is_deterministic():
    # Test that multiple calls with the same input give the same output
    pos = torch.tensor([0.0, 1.0, 2.0])
    embed_dim = 4
    codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos); emb1 = codeflash_output # 109μs -> 91.8μs (19.8% faster)
    codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos); emb2 = codeflash_output # 38.7μs -> 27.9μs (38.8% faster)
    torch.testing.assert_close(emb1, emb2)

def test_output_range():
    # Test that all outputs are in [-1, 1]
    pos = torch.linspace(-100, 100, 50)
    embed_dim = 10
    codeflash_output = get_1d_sincos_pos_embed_from_grid_torch(embed_dim, pos); emb = codeflash_output # 104μs -> 85.4μs (22.3% faster)
# codeflash_output is used to check that the output of the original code is the same as that of the optimized code.
```

</details>


To edit these changes `git checkout codeflash/optimize-get_1d_sincos_pos_embed_from_grid_torch-mcw331xg` and push.

[![Codeflash](https://img.shields.io/badge/Optimized%20with-Codeflash-yellow?style=flat&color=%23ffc428&logo=data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iNDgwIiBoZWlnaHQ9ImF1dG8iIHZpZXdCb3g9IjAgMCA0ODAgMjgwIiBmaWxsPSJub25lIiB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciPgo8cGF0aCBmaWxsLXJ1bGU9ImV2ZW5vZGQiIGNsaXAtcnVsZT0iZXZlbm9kZCIgZD0iTTI4Ni43IDAuMzc4NDE4SDIwMS43NTFMNTAuOTAxIDE0OC45MTFIMTM1Ljg1MUwwLjk2MDkzOCAyODEuOTk5SDk1LjQzNTJMMjgyLjMyNCA4OS45NjE2SDE5Ni4zNDVMMjg2LjcgMC4zNzg0MThaIiBmaWxsPSIjRkZDMDQzIi8+CjxwYXRoIGZpbGwtcnVsZT0iZXZlbm9kZCIgY2xpcC1ydWxlPSJldmVub2RkIiBkPSJNMzExLjYwNyAwLjM3ODkwNkwyNTguNTc4IDU0Ljk1MjZIMzc5LjU2N0w0MzIuMzM5IDAuMzc4OTA2SDMxMS42MDdaIiBmaWxsPSIjMEIwQTBBIi8+CjxwYXRoIGZpbGwtcnVsZT0iZXZlbm9kZCIgY2xpcC1ydWxlPSJldmVub2RkIiBkPSJNMzA5LjU0NyA4OS45NjAxTDI1Ni41MTggMTQ0LjI3NkgzNzcuNTA2TDQzMC4wMjEgODkuNzAyNkgzMDkuNTQ3Vjg5Ljk2MDFaIiBmaWxsPSIjMEIwQTBBIi8+CjxwYXRoIGZpbGwtcnVsZT0iZXZlbm9kZCIgY2xpcC1ydWxlPSJldmVub2RkIiBkPSJNMjQyLjg3MyAxNjQuNjZMMTg5Ljg0NCAyMTkuMjM0SDMxMC44MzNMMzYzLjM0NyAxNjQuNjZIMjQyLjg3M1oiIGZpbGw9IiMwQjBBMEEiLz4KPC9zdmc+Cg==)](https://codeflash.ai)