@LeWaldm 2025-10-31
Hi @adamjstewart , thank you for the reminder! I checked the commits, this is functionally equivalent and a bit faster. We can go ahead and merge this!